+++ app/components/dashboard/dashboard-service.js
@@ -1,0 +1,102 @@
+'use strict';
+
+angular.module('dashboard')
+.service('DashboardService', [ function () {
+
+  this.GRAPH_PADDING = 13; // Padding around the graph svg. Not to be confused
+                          // with the padding inside the svg which is used for
+                          // axis and labels.
+  var ROW_BOTTOM_MARGIN = 10; // Pixels between graph rows.
+
+
+  /**
+   * Combines timeseries, with other chartable active data to dashboard data.
+   *
+   * Graphs are an array of graph-directive objects representing timeseries
+   * and raster data.
+   * {
+   *   'type': 'type',
+   *   content: [{
+   *     data: [],
+   *     keys: {},
+   *     labels: {},
+   *     color: ''
+   *   }]
+   * }
+   *
+   * @param  {array} timeseries Data source timeseries from timseriesService.
+   * @param  {array} assets     Data source DataService.assets.
+   * @param  {array} geometries Data source DataService.geometries.
+   * @return {array} graph
+   */
+  this.buildGraphs = function (timeseries, assets, geometries) {
+
+    var graphs = [];
+
+    if (timeseries.length) {
+      var content = [];
+      timeseries.forEach(function (ts) {
+        content.push(ts);
+      });
+      graphs.push({ 'type': 'temporalLine', 'content': content });
+    }
+
+    assets.forEach(function (asset) {
+      graphs = addPropertyData(graphs, asset.properties);
+    });
+
+    geometries.forEach(function (geometry) {
+      graphs = addPropertyData(graphs, geometry.properties);
+    });
+
+    return graphs;
+  };
+
+  /**
+   * Creates a dimensions object for graph-directive.
+   *
+   * @param  {angular element} element   element to draw graphs in.
+   * @param  {[int}            nGraphs   number of graphs in dashboard.
+   * @param  {boolean}         showXAxis should be true for non temporal graphs.
+   * @return {object}          dimension object per graph.
+   */
+  this.getDimensions = function (element, nGraphs, showXAxis) {
+
+    return {
+      width: element.width() - this.GRAPH_PADDING,
+      height: getGraphHeight(element, nGraphs),
+      padding: {
+        top: 10,
+        right: 10,
+        bottom: showXAxis ? 40 : 15,
+        left: 10
+      }
+    };
+  };
+
+  /**
+   * Adds DataService.assets|geometries.properties to dashboard graphs object.
+   *
+   * @param {object} graphs     dashboard graph object.
+   * @param {object} properties asset or geometries properties.
+   */
+  var addPropertyData = function (graphs, properties) {
+    _.forEach(properties, function (property, slug) {
+      if (property.data.length > 1) {
+        var item = {
+          data: property.data,
+          keys: {x: 0, y: {y0: 1, y1: 1}},
+          labels: {x: 'm', y: property.unit }
+        };
+        graphs.push({ type: 'distance', content: [item] });
+      }
+    });
+    return graphs;
+  };
+
+  var getGraphHeight = function (element, nGraphs) {
+    return (element.height() - ROW_BOTTOM_MARGIN * nGraphs) / nGraphs;
+  };
+
+
+}]);