+++ app/components/timeseries/timeseries-service.js
@@ -2,29 +2,59 @@
  * Service to handle timeseries retrieval.
  */
 angular.module('timeseries')
-  .service("TimeseriesService", ['CabinetService', '$q',
-    function (CabinetService, $q) {
+  .service("TimeseriesService", ['$q', 'State', '$http', 'notie',
+    function ($q, State, $http, notie) {
+
+    var GRAPH_WIDTH = 320; // Width of drawing area of box graphs.
+
+    // Contains timeseries metadata and data as comes from api. It mirrors
+    // State.seletected.timeseries.
+    this.timeseries = [];
+
+    this.minPoints = GRAPH_WIDTH; // default
+
+    var service = this;
+
+    var _timeseries = [];
+    Object.defineProperty(State.selected, 'timeseries', {
+      get: function () { return _timeseries; },
+      set: function (timeseries) {
+        console.log('State.selected.timeseries:', timeseries);
+        service.syncTime(timeseries);
+        _timeseries = timeseries;
+      }
+    });
+
+    this.syncTime = function (timeseries) {
+      var promise = {};
+
+      // Return empty timeseries when empty selection
+      if (timeseries && timeseries.length === 0) {
+        var defer = $q.defer();
+        defer.resolve([]);
+        promise = defer.promise;
+      }
+
+      else {
+        promise = this._getTimeseries(timeseries || State.selected.timeseries, State.temporal, service.minPoints);
+      }
+
+      promise.then(function (ts) {
+        service.timeseries = ts;
+        console.log('TimeseriesService.timeseries:', service.timeseries);
+      })
+
+      .then(function () {
+
+          if (service.onTimeseriesChange) {
+            service.onTimeseriesChange();
+          }
+
+      });
+    };
 
     var localPromises = {};
 
-    this._getTimeseries = function (id, timeState, minPoints, aggWindow) {
-      // Cancel consecutive calls for the same ts.
-      if (localPromises[id]) {
-        localPromises[id].reject('consecutive');
-      }
-      localPromises[id] = $q.defer();
-      var params = {
-        object: id,
-        start: timeState.start ? parseInt(timeState.start, 10): undefined,
-        end: timeState.end ? parseInt(timeState.end, 10): undefined,
-      };
-
-      minPoints ? params.min_points = minPoints : params.window = aggWindow;
-
-      return CabinetService.timeseries.get(params, {
-        timeout: localPromises[id].promise
-      });
-    };
 
     /**
      * @function
@@ -41,84 +71,140 @@ angular.module('timeseries')
      *                        aggregated to a sensible size.
      *
      */
-    this.getTimeSeriesForObject = function (
-        objectId,
-        start,
-        end,
-        minPoints,
-        aggWindow,
-        defer
-      ) {
+    this._getTimeseries = function (uuids, timeState, minPoints) {
+      // Cancel consecutive calls for the same ts.
+      var id = uuids.join(',');
+      if (localPromises[id]) {
+        localPromises[id].reject('consecutive');
+      }
+      localPromises[id] = $q.defer();
+      var params = {
+        uuid: id,
+        start: timeState.start ? parseInt(timeState.start, 10): undefined,
+        end: timeState.end ? parseInt(timeState.end, 10): undefined,
+      };
+
+      minPoints ? params.min_points = minPoints : params.window = timeState.aggWindow;
+
+      return $http({
+        url: 'api/v2/timeseries/',
+        method: 'GET',
+        params: params
+      })
+
+      .then(function (response) {
+
+        return response.data;
+
+      }, errorFn)
+
+      .then(filterTimeseries, null)
+      .then(formatTimeseriesForGraph, null);
+    };
+
+    var errorFn = function (err) {
+      if (err.status === 420) {
+        // Cancel normal operations
+        return $q.reject(err);
+      }
+      else if (err.status >= 500 && err.status < 600) {
+        notie.alert(3, 'Lizard encountered a problem retrieving your timeseries.', 3);
+        // Cancel normal operations
+        return $q.reject(err);
+      }
+      window.Raven.captureException(err);
+      return err; // continue anyway
+    };
+
+    this.getTimeSeriesForAsset = function (asset) {
+      // Get ts if asset has ts at the back but not yet at the front.
+      if (asset.num_timeseries && asset.timeseries === undefined) {
+        return $http({
+          url: 'api/v2/timeseries/',
+          method: 'GET',
+          params: { object: asset.entity_name + '$' + asset.id }
+        })
+
+        .then(function (response) {
+          if (response.data.results.length) {
+            asset.timeseries = response.data.results;
+          }
+          return asset;
+        });
+      }
+
+      else {
+        var defer = $q.defer();
+        defer.resolve(asset);
+        return defer.promise;
+      }
+
+    };
+
+    var formatTimeseriesForGraph = function (timeseries) {
+
+      var graphTimeseriesTemplate = {
+        id: '', //uuid
+        data: [],
+        labels: {
+          x: '',
+          y: ''
+        },
+        keys: { x: 'timestamp', y: { y0: 'min', y1: 'max' } }
+      };
+
+      var result = [];
+      timeseries.forEach(function (ts) {
+        var graphTimeseries = angular.copy(graphTimeseriesTemplate);
+        graphTimeseries.data = ts.events;
+        graphTimeseries.id = ts.uuid;
+        result.push(graphTimeseries);
+      });
+      return result;
+
+    };
+
+    var filterTimeseries = function (response) {
 
       // maximum number of timeseries events, more probably results in a
       // memory error.
       var MAX_NR_TIMESERIES_EVENTS = 25000;
-      var promise = this._getTimeseries(
-        objectId,
-        {
-          start: start,
-          end: end
-        },
-        minPoints,
-        aggWindow
-      ).then(function (response) {
 
-        if (!response) { return; }
+      var filteredResult = [];
 
-         // Filter out the timeseries with too little measurements. And ts
-         // without parameter unit info.
-        var filteredResult = [];
-        angular.forEach(response.results, function (ts) {
-          var msg = '';
-          if (ts.events === null) {
-            filteredResult.push(ts);
-          } else if (ts.events.length > 1 &&
-              ts.events.length < MAX_NR_TIMESERIES_EVENTS) {
+      angular.forEach(response.results, function (ts) {
+        var msg = '';
+        if (ts.events === null) {
+          filteredResult.push(ts);
+        } else if (ts.events.length < MAX_NR_TIMESERIES_EVENTS) {
 
-            if (ts.parameter_referenced_unit === null) {
-              ts.parameter_referenced_unit = {};
-            }
-
-            filteredResult.push(ts);
-
-          // Else: output a message to the console and an error to sentry.
-          } else if (ts.events.length > MAX_NR_TIMESERIES_EVENTS) {
-            msg = 'Timeseries: '
-              + ts.uuid
-              + ' has: '
-              + ts.events.length
-              + ' events, while '
-              + MAX_NR_TIMESERIES_EVENTS
-              + ' is the maximum supported amount';
-            window.Raven.captureException(new Error(msg));
-            console.info(msg);
-          } else if (!ts.parameter_referenced_unit) {
-            msg = 'Timeseries: '
-              + ts.uuid
-              + ' has no valid parameter_referenced_unit';
-            window.Raven.captureException(new Error(msg));
-            console.info(msg);
+          if (ts.parameter_referenced_unit === null) {
+            ts.parameter_referenced_unit = {};
           }
-        });
 
-        // Legacy dataservice calls this function with a defer which is not used
-        // by the timeseries directive.
-        //
-        // TODO: when we refactored the dashboard/time-ctx we can strip
-        // timeseries from the data-service and remove this.
-        if (defer) {
-          defer.notify({
-            data: filteredResult,
-            layerGroupSlug: 'timeseries',
-            layerSlug: 'timeseries',
-          });
+          filteredResult.push(ts);
+
+        // Else: output a message to the console and an error to sentry.
+        } else if (ts.events.length > MAX_NR_TIMESERIES_EVENTS) {
+          msg = 'Timeseries: '
+            + ts.uuid
+            + ' has: '
+            + ts.events.length
+            + ' events, while '
+            + MAX_NR_TIMESERIES_EVENTS
+            + ' is the maximum supported amount';
+          window.Raven.captureException(new Error(msg));
+          console.info(msg);
+        } else if (!ts.parameter_referenced_unit) {
+          msg = 'Timeseries: '
+            + ts.uuid
+            + ' has no valid parameter_referenced_unit';
+          window.Raven.captureException(new Error(msg));
+          console.info(msg);
         }
-
-        response.results = filteredResult;
-        return response; // accomadate chaining.
       });
 
-      return promise;
+      return filteredResult;
     };
 
   }