+++ CHANGES.rst
@@ -4,6 +4,16 @@ Changelog of lizard-nxt client
 Unreleased (2.6.2) (XXXX-XX-XX)
 -------------------------------
 
+- Draw selected timeseries and raster data in dashboard.
+
+- No longer load map before dashboard. Dashboard no longer needs map.
+
+- Refactored timeseries. State.selected.timeseries keeps track of selected ts
+  TimeseriesService synchronizes the data with the selection. Draw graphs for
+  TimeseriesService.timeseries.
+
+- Only startPlugins if Lizard plugins is loaded.
+
 - Add chalk to npm dependencies.
 
 - Refine the data-menu restyle.